+++ libs/common/src/lib/exceptions/index.ts
@@ -1,0 +1,3 @@
+import { NotFoundError } from './not-found-error';
+
+export { NotFoundError };