+++ libs/common/src/lib/interfaces/benchmark-property.interface.ts
@@ -1,0 +1,3 @@
+export interface BenchmarkProperty {
+  symbolProfileId: string;
+}